pyproject.toml
[project]
name = "multi-agent-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "dodopayments>=1.0",
    "redis>=5.0",
    "rq>=1.16",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    "pygithub>=2.0",
]
